[YS-416] refactor: 회원 탈퇴 버튼 컴포넌트 분리
